Estimate PCI hole size before starting QEMU
Add a XenStore node with the expected top of RAM address below the 32b
address range limit. dm-agent will then pass it to QEMU.

The estimation is made based on 3 assumptions:
- Default QEMU configuration does not require more than 32MB PCI hole
- The PCI hole cannot grow bigger than 2GB
- The "default" PCI hole is 0xf0000000-0xfc000000 (see xen e820.h)

Hvmloader will re-process that value. It is unclear if we should let
that happen or patch it. For now we should always find the same value
and stay consistent.

OXT-243

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>