ocamlfind: Use across the board for libraries
ocamlfind is a frontend with the package manager.
All libraries in that repository use it to install/remove built artefacts,
except json that uses it to find where camomile is and get the options down to
ocamlc/ocamlopt (ocamlfind <ocamlc|ocamlopt> -package camomile). It does this
by setting OCAMLC/OCAMLTOP which are already set to pass the cross environment
options.

In order to avoid re-writing more build-system, use ocamlfind for every library
to invoke the byte-code compiler and native compiler and pass additional
package (-package) options through OCAMLCOPT and OCAMLOPTOPT. This way OCAMLC
and OCAMLOPT can define the specific cross environment and ocamlfind can
provide its services to include packages through findlib.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

OXT-1787